Update lodash and dev dependencies
Version 3 of the Lodash module contains a security vulnerability. That
vulnerability was patched in the 4.17.5 release.

Note that version 4 of Lodash removes support for Internet Explorer 6,
7, and 8. The require-folder-tree project does not formally document
compatibility with web browsers, but the source is currently dependent
on Array prototype methods introduced in ES5. This means that updating
Lodash does not effect the set of browsers supported by this project.

[1] https://hackerone.com/reports/310443
[2] https://github.com/lodash/lodash/wiki/Changelog#v400